requirement.txt
line-bot-sdk==3.7
flask==3.0.0
python-dotenv==1.0.1
pandas==2.2.2
google-generativeai==0.8.4